feat: `auto_close` (#29)
By default, FTerm closes the terminal buffer as soon as the shell or command exists and there is no way to override this behavior.

This PR provides an option i.e. auto_close to override that behavior and prevents buffer from closing. FWIW this is also the neovim's default behavior for terminals.