change: make `setup()` optional (#34)
After removing VimResized autocommand, setup() is not doing anything other than creating the floating window configuration. So, why not make the setup optional by providing the default configuration as it is.

The only concern is if we need something magical ahead of time like VimResized then we might need to make setup() a requirement.