Support more window options
This plugin is perfect for my needs but the one issue I had was that I
wanted my terminal to be in the bottom right of the screen. I know I
could calculate the right x/y to pass in, but in this instance I knew I
wanted it anchored to the bottom right, and that I should use the number
of cols/rows for the values.

Therefore I've created this PR to add the flexibility to let a user pass
in the `anchor` option, which maps right through to the nvim_open_win
property, and also pass `col` and `row` into `dimensions`, which will be
used above `x` and `y`.

I'm very happy to change the approach or API of these options, if it's
something you're willing to land, just let me know. Equally no problem
if you don't want to merge this; I'll probably keep my own fork lying
around so I can use it myself. Thanks!